Leetcode 771 solution in Python
The first solution does it in O(s+j) where s and j are the length of the strings. By changing it to a set or hashset we can makeup the lookup more efficient and hence the time complexity changes to O(s*j). The logic is otherwise easy to implement